[WIP]security/intel: Add option to eanble SMM flash access only
On platforms where the boot media can be updated externally, for e.g.
using a BMC, add the possibility to enable writes in SMM only.

This allows to protect the BIOS region even without the use of vboot, but
keeps the SMMSTORE working for the use in payloads.

UNTESTED.

B: Removed pre-PCH platforms; unable to test

Change-Id: I157db885b5f1d0f74009ede6fb2342b20d9429fa
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>