drivers/generic/cbfs-serial: Update driver to read UUID from CBFS
Modify the existing cbfs-serial driver to also read the UUID from
a text file in CBFS, just as a serial number is done today. When
driver is selected and a system_uuid exists in the CBFS the UUID
is read from the file and populated in the SMBIOS table. Note
that in order to work there must be no newline character at the
end of the file, this is also true of the serial_number.

Tested on a Google Tricky (Dell 3010 Chromebox).

Change-Id: I140a61670cec1318c095e18815c6bc7c1bf78cd6
Signed-off-by: Chris Morgan <macromorgan@hotmail.com>