soc/intel/*: Add TPM decode range to soc_get_fixed_mmio_ranges()
The TPM memory MMIO range is always decoded by the PCH, regardless
if it's LPC, SPI or fTPM.

Add the memory range to fixed MMIO ranges list to prevent LGMR register
being written with this address. Reserved 32 KiB as that seems to be the
minimal common reserved size for the TPM MMIO area.

B: Removed platforms other than Skylake; unable to test. Changed length to
0x5000 from 0x8000, following TCG PTP and TIS specifications.

Change-Id: I8f2257b55a712b936763cfd289c2c4b1633e8049
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>